polls/requirements.txt
Django==3.0.4
djangorestframework==3.7.7